crud category dan produk

+++ app/Http/Controllers/CategoryController.php
@@ -66,9 +66,9 @@ class CategoryController extends Controller
         //ADAPUN withCount() SERUPA DENGAN EAGER LOADING YANG MENGGUNAKAN with()
         //HANYA SAJA withCount() RETURNNYA ADALAH INTEGER
         //JADI NNTI HASIL QUERYNYA AKAN MENAMBAHKAN FIELD BARU BERNAMA child_count YANG BERISI JUMLAH DATA ANAK KATEGORI
-        $category = Category::withCount(['child'])->find($id);
+        $category = Category::withCount(['child', 'product'])->find($id);
 
-        if ($category->child_count == 0) {
+        if ($category->child_count == 0  && $category->product_count == 0) {
             $category->delete();
             return redirect(route('category.index'))->with(['success' => 'Kategori dihapus!']);
         }
